Backlog #54/#55: rozet sistemi + haftalık performans özeti
İçerideki rekabeti artırmak için iki retention mekaniği: (1) rozet
sistemi — ilk gol, hat-trick, 5 maçlık seri, güvenilir oyuncu, yıldız
reyting rozetleri; skor/istatistik/reyting girişleri sonrası otomatik
kontrol edilir, kazanılınca akışa "rozet kazandı" kartı düşer. (2)
haftalık performans özeti — Spotify Wrapped mantığı, her Pazar en az
1 maça katılan oyunculara otomatik maç/gol/asist/puan özeti gönderisi.
Her ikisi de mevcut auto-post ve scheduled-command desenlerini izliyor.

Ayrıca daha önce hiçbir gerçek MySQL veritabanında çalışmamış olan
districts migration'ındaki FK tip uyuşmazlığı (city_id bigint vs
cities.id smallint) düzeltildi.

+++ api/database/migrations/2026_07_11_230000_create_districts_table.php
@@ -12,7 +12,10 @@ return new class extends Migration
     {
         Schema::create('districts', function (Blueprint $Table) {
             $Table->id();
-            $Table->unsignedBigInteger('city_id');
+            // cities.id plaka kodu (unsignedSmallInteger) — FK tipi eşleşmeli,
+            // aksi halde MySQL 8 "incompatible" hatası verir (SQLite'ta bu
+            // uyuşmazlık sessizce geçiyordu, testler bu yüzden yakalamamıştı).
+            $Table->unsignedSmallInteger('city_id');
             $Table->string('name', 60);
             $Table->foreign('city_id')->references('id')->on('cities')->cascadeOnDelete();
             $Table->unique(['city_id', 'name']);
